Native transaction list and configurable clickable view (#325)
I think, that the current implementation of the list is production-ready. I've made some testing on the emulator and it feels way more robust than the initial JS implementation. Hope it will meet your expectations. Aside from that, in the last few commits, you may find a NativeButton component which is nothing other than a ButtonPressAnimation component re-implemented in Swift (as we discussed with @osdnk).

I would love to receive your feedback and adjust details for the list and/or button if needed. There are some tiny things that can be improved (code-wise), but they shouldn't anyhow interfere with the business logic of the app, so the PR is safe to merge.