Set focused as a param only in native stack (#424)
observed that this commit 1d5a130
made a regression. "Normal" stack animation is no longer visible when transitioning out (with no gesture)

According to the purpose of this commit, there's no reason for having this logic when routes are not inside the native stack. I' adding a proper check for this.