Restructure navigators (#619)
I realized that the Slack Bottom sheet is not working on iOS 12. I was thinking firstly that's this is some issue with native code, but then realized that the problem was with the structure of navigators I've built. Long story short: bottom sheet navigator was inside the native stack and it's not available under iOS 13.
Ofc, the Slack Bottom Sheet works well on iOS 12 so I just moved this navigator on the top of other navigators.

While doing it I realized that the structure of our navigation is not very complicated but because it's all included in one file w/o any abstraction it's quite difficult to read and debug. So I decided to build a small abstraction over our navigation structure. I moved routes' names to the other file and made them constants (it might help us while refactoring and prevent bugs like typos in names). I moved our logic for sending analytics and adjusting the status bar. Also, I've added a few helpers to follow the DRY rule as much as it's possible. Then I managed to reduce Routes.io.js +/- twice and similarly on Android.

Then I replaced all navigate('name') to navigate(CONSTANT_NAME)

I enhanced the settings of our eslint to not throw on ios, android files. Then removed eslint-ignores